+++ axolotl-api/src/world_gen/noise/density/mod.rs
@@ -38,11 +38,11 @@ impl From<&'static str> for BuildDefResult {
     }
 }
 pub trait DensityContext {
-    fn get_x(&self) -> i64;
+    fn get_x(&self) -> i32;
 
     fn get_y(&self) -> i16;
 
-    fn get_z(&self) -> i64;
+    fn get_z(&self) -> i32;
 }
 /// The Current Density State
